add management
git-svn-id: http://192.168.1.50/svn/bhp_crypto@79 dc6f6662-d1d5-42af-b45d-bfa79074b674

+++ utils/setup_new_keys.py
@@ -3,6 +3,7 @@ __authors__ = [
 ]
 
 import os
+import sys
 from datetime import datetime
 from bhp_crypto.classes import Crypter
 
@@ -10,9 +11,8 @@ from bhp_crypto.classes import Crypter
 def setup_new_keys():
 
     """ Utility to generate all new keys for the project."""
-
     datestring = datetime.today().strftime('%Y%m%d%H%M%S%f')
-    crypter = Crypter()
+    crypter = Crypter(no_preload=True)
     filenames = []
     for algorithm, value in crypter.VALID_MODES.iteritems():
         if not isinstance(value, dict):
@@ -36,12 +36,28 @@ def setup_new_keys():
             oldpath = os.path.join(os.path.realpath('.'), filename)
             newpath = os.path.join(os.path.join(
                                        os.path.realpath('.'), path), filename)
-            print filename
-            os.rename(oldpath, newpath)
-        except:
+            if os.path.exists(oldpath):
+                os.rename(oldpath, newpath)
+                print 'copied {0}'.format(filename)
+        except OSError as e:
             print ('Failed to copy {0} to backup '
                   'folder {1}'.format(filename, path))
-    crypter = Crypter()
-    crypter.create_new_rsa_key_pairs(suffix='')
-    crypter.create_new_salt(suffix='')
-    crypter.create_aes_key(suffix='')
+            break
+    # confirm target folder has no keys
+    old_keys_exist = False
+    for filename in filenames:
+        oldpath = os.path.join(os.path.realpath('.'), filename)
+        if os.path.exists(oldpath):
+            old_keys_exist = True
+            break
+
+    if old_keys_exist:
+        print 'Failing. Old keys are still in the target folder. Try moving them manually.'
+    else:
+        print 'Creating new keys'
+        del crypter
+        crypter = Crypter(no_preload=True)
+        crypter.create_new_rsa_key_pairs(suffix='')
+        crypter.create_new_salt(suffix='')
+        crypter.create_aes_key(suffix='')
+        sys.stdout.flush()
